// Frontend/src/app/models/bookShop.ts
import {Book} from "./book";

export class BookShop {
  id?: any;
  title?: string;
  shop_name?:string;
  shop_number?: number;
  location?: string;
  books?:Book[];
  contact_no?:number;
  email?: string;
  // genre?: string[];
  // registration?: Date;
}

// @Id
// private  String id;
// private String title;
// private String shop_name;
// private String shop_number;
// private String location;
// private Book[] books;
// private int contact_no;
// private String email;